+++ editor/include/adminer.inc.php
@@ -119,6 +119,9 @@ class Adminer {
 
 	function login(string $login, string $password) {
 		// the same as in Adminer, a customization replacing the database credentials by its own overrides this method anyway
+		if ($password != "" && !Driver::$passwords) {
+			return lang('The database does not support passwords.');
+		}
 		if ($password == "" || !password_required()) { // !password_required() - the server would accept any password
 			return lang('Adminer does not support accessing a database without a password (<a href="https://www.adminer.org/en/password/"%s>more information</a>).', target_blank());
 		}